9 ошибок

+++ client/src/redux/slices/groups/thunkActions.ts
@@ -1,5 +1,4 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import GroupServise from '../../../services/groups';
 import teacherGroupService from '../../../services/teacherGroups';
 import AdminService from '../../../services/adminService';
 import type{ TeacherGroupType } from '../../../types/admin';
@@ -8,6 +7,11 @@ export const thunkTeacherGroups = createAsyncThunk('groupsSlise/thunkTeacherGrou
   teacherGroupService.getTeacherGroups(),
 );
 
-export const thunkDeleteTeacher = createAsyncThunk('groupsSlise/thunkDeleteTeacher', async (teacherToDelete: TeacherGroupType | null) =>
-  AdminService.deleteTeacher(teacherToDelete),
-);
+export const thunkDeleteTeacher = createAsyncThunk('groupsSlice/thunkDeleteTeacher', async (teacherToDelete: TeacherGroupType | null) => {
+  if (teacherToDelete !== null) {
+    // Proceed with the delete operation
+    return AdminService.deleteTeacher(teacherToDelete);
+  } 
+    return null;
+  
+});
